debugai/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "debugai"
version = "1.0.0"
description = "AI-Powered Log Analysis & Debugging CLI - Reduce debugging time by 60-75%"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "DebugAI Team", email = "team@debugai.dev" }
]
keywords = [
    "debugging",
    "logging",
    "ai",
    "cli",
    "developer-tools",
    "log-analysis",
    "error-tracking",
    "gemini",
    "devops"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Debuggers",
    "Topic :: System :: Logging",
    "Topic :: Utilities",
]

dependencies = [
    # CLI Framework
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "click>=8.1.7",
    
    # AI & ML
    "google-generativeai>=0.3.0",
    
    # Log Parsing
    "python-dateutil>=2.8.2",
    "regex>=2023.12.25",
    
    # Data Processing
    "pandas>=2.1.4",
    "numpy>=1.26.3",
    "cachetools>=4.2.1",
    
    # Storage
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",
    
    # Configuration
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "toml>=0.10.2",
    
    # Docker Integration
    "docker>=7.0.0",
    
    # HTTP & Async
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    
    # Utilities
    "watchdog>=3.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.8.0",
    "ruff>=0.1.11",
    "black>=23.12.1",
    "pre-commit>=3.6.0",
    "ipython>=8.20.0",
]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.3",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
debugai = "debugai.cli.main:app"

[project.urls]
Homepage = "https://github.com/debugai/debugai"
Documentation = "https://debugai.dev/docs"
Repository = "https://github.com/debugai/debugai"
Issues = "https://github.com/debugai/debugai/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/debugai"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict
]

[tool.ruff.isort]
known-first-party = ["debugai"]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--cov=src/debugai",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["src/debugai"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
